requirements/benchmarking.txt
# Generated via tools/generate_requirements.py and pre-commit hook.
# Do not edit this file; modify pyproject.toml instead.
asv
virtualenv
pytest-codspeed==4.2.0
